chore: rename the css
