+++ packages/create-shvui/templates/typescript/src/routes/module.css
